refactor: extract page renderers and unify LLM providers (Phase 1+2)
Phase 1 - God-module reduction in main.py:
- Extract 742-line _build_topic_page() → topic_page_generator.py
- Extract 659-line _build_media_page() → new media_page_generator.py
- Replace both methods with thin stubs delegating to standalone modules
- main.py reduced from 3013 → ~1221 lines

Phase 2 - LLM provider consolidation:
- Create llm_client.py with ProviderSpec dataclass and
  call_openai_compatible() handling retry, rate-limit, and response parsing
  uniformly for all four OpenAI-compatible providers (Groq, OpenRouter,
  OpenCode, Mistral)
- Replace four 60-80 line fat methods in editorial_generator.py with
  four 6-line thin wrappers; eliminates ~280 LOC of duplicated HTTP logic

Previously committed (a87ac3a): security fixes, json_utils.py, enrich_content.py
