+++ scripts/editorial_generator.py
@@ -22,6 +22,13 @@ from pathlib import Path
 from typing import Dict, List, Optional
 
 from json_utils import escape_control_chars_in_strings
+from llm_client import (
+    call_openai_compatible,
+    GROQ_SPEC,
+    OPENROUTER_SPEC,
+    OPENCODE_SPEC,
+    MISTRAL_SPEC,
+)
 
 
 _SAFE_URL_SCHEMES = {"http", "https", "mailto"}
@@ -1799,256 +1806,36 @@ DATE: {datetime.now().strftime('%B %d, %Y')}"""
         self, prompt: str, max_tokens: int = 800, max_retries: int = 1
     ) -> Optional[str]:
         """Call OpenRouter API with free models (primary)."""
-        if not self.openrouter_key:
-            logger.warning("No OpenRouter API key available")
-            return None
-
-        # Check rate limits before calling
-        rate_limiter = get_rate_limiter()
-        status = check_before_call("openrouter")
-
-        if not status.is_available:
-            logger.warning(f"OpenRouter not available: {status.error}")
-            return None
-
-        if status.wait_seconds > 0:
-            logger.info(
-                f"Waiting {status.wait_seconds:.1f}s for OpenRouter rate limit..."
-            )
-            time.sleep(status.wait_seconds)
-
-        # Free models to try in order of preference
-        free_models = [
-            "meta-llama/llama-3.3-70b-instruct:free",
-            "deepseek/deepseek-r1-0528:free",
-            "google/gemma-3-27b-it:free",
-        ]
-
-        for model in free_models:
-            for attempt in range(max_retries):
-                try:
-                    logger.info(
-                        f"Trying OpenRouter {model} (attempt {attempt + 1}/{max_retries})"
-                    )
-                    response = self.session.post(
-                        "https://openrouter.ai/api/v1/chat/completions",
-                        headers={
-                            "Authorization": f"Bearer {self.openrouter_key}",
-                            "Content-Type": "application/json",
-                            "HTTP-Referer": "https://dailytrending.info",
-                            "X-Title": "DailyTrending.info",
-                        },
-                        json={
-                            "model": model,
-                            "messages": [{"role": "user", "content": prompt}],
-                            "max_tokens": max_tokens,
-                            "temperature": 0.7,
-                        },
-                        timeout=60,
-                    )
-                    response.raise_for_status()
-
-                    # Update rate limiter from response headers
-                    rate_limiter.update_from_response_headers(
-                        "openrouter", dict(response.headers)
-                    )
-
-                    result = (
-                        response.json()
-                        .get("choices", [{}])[0]
-                        .get("message", {})
-                        .get("content")
-                    )
-                    if result:
-                        logger.info(f"OpenRouter success with {model}")
-                        return result
-                except requests.exceptions.HTTPError as e:
-                    if response.status_code == 429:
-                        # Parse retry-after header if available
-                        retry_after = response.headers.get("Retry-After", "10")
-                        try:
-                            wait_time = min(float(retry_after), self.MAX_RETRY_WAIT)
-                        except ValueError:
-                            wait_time = self.MAX_RETRY_WAIT
-                        logger.warning(
-                            f"OpenRouter {model} rate limited, waiting {wait_time}s (attempt {attempt + 1}/{max_retries})"
-                        )
-                        time.sleep(wait_time)
-                        continue
-                    logger.warning(f"OpenRouter {model} failed: {e}")
-                    break  # Try next model
-                except Exception as e:
-                    logger.warning(f"OpenRouter {model} failed: {e}")
-                    break  # Try next model
-
-        logger.warning("All OpenRouter models failed")
-        return None
+        result = call_openai_compatible(
+            OPENROUTER_SPEC, self.openrouter_key, prompt, max_tokens, max_retries,
+            self.session, self.MAX_RETRY_WAIT,
+        )
+        return result
 
     def _call_groq_direct(
         self, prompt: str, max_tokens: int = 800, max_retries: int = 1
     ) -> Optional[str]:
         """Call Groq API directly (fallback)."""
-        if not self.groq_key:
-            return None
-
-        # Check rate limits before calling
-        rate_limiter = get_rate_limiter()
-        status = check_before_call("groq")
-
-        if not status.is_available:
-            logger.warning(f"Groq not available: {status.error}")
-            return None
-
-        if status.wait_seconds > 0:
-            logger.info(f"Waiting {status.wait_seconds:.1f}s for Groq rate limit...")
-            time.sleep(status.wait_seconds)
-
-        # Proactive rate limiting
-        elapsed = time.time() - self._last_call_time
-        if elapsed < self.MIN_CALL_INTERVAL:
-            time.sleep(self.MIN_CALL_INTERVAL - elapsed)
-
-        for attempt in range(max_retries):
-            try:
-                self._last_call_time = time.time()
-                response = self.session.post(
-                    "https://api.groq.com/openai/v1/chat/completions",
-                    headers={
-                        "Authorization": f"Bearer {self.groq_key}",
-                        "Content-Type": "application/json",
-                    },
-                    json={
-                        "model": "llama-3.3-70b-versatile",
-                        "messages": [{"role": "user", "content": prompt}],
-                        "max_tokens": max_tokens,
-                        "temperature": 0.7,
-                    },
-                    timeout=60,
-                )
-                response.raise_for_status()
-
-                # Update rate limiter from response headers
-                rate_limiter.update_from_response_headers(
-                    "groq", dict(response.headers)
-                )
-
-                return (
-                    response.json()
-                    .get("choices", [{}])[0]
-                    .get("message", {})
-                    .get("content")
-                )
-            except requests.exceptions.HTTPError as e:
-                if response.status_code == 429:
-                    # Parse retry-after header if available
-                    retry_after = response.headers.get("Retry-After", "10")
-                    try:
-                        wait_time = min(float(retry_after), self.MAX_RETRY_WAIT)
-                    except ValueError:
-                        wait_time = self.MAX_RETRY_WAIT
-                    logger.warning(
-                        f"Groq rate limited, waiting {wait_time}s (attempt {attempt + 1}/{max_retries})"
-                    )
-                    time.sleep(wait_time)
-                    continue
-                logger.error(f"Groq API error: {e}")
-                return None
-            except Exception as e:
-                logger.error(f"Groq API error: {e}")
-                return None
-
-        logger.warning("Groq API: Max retries exceeded")
-        return None
+        timing = [self._last_call_time]
+        result = call_openai_compatible(
+            GROQ_SPEC, self.groq_key, prompt, max_tokens, max_retries,
+            self.session, self.MAX_RETRY_WAIT, self.MIN_CALL_INTERVAL, timing,
+        )
+        self._last_call_time = timing[0]
+        return result
 
     def _call_opencode(
         self, prompt: str, max_tokens: int = 800, max_retries: int = 1
     ) -> Optional[str]:
         """Call OpenCode API with free models (glm-4.7-free, minimax-m2.1-free)."""
         opencode_key = os.getenv("OPENCODE_API_KEY")
-        if not opencode_key:
-            return None
-
-        # Check rate limits before calling
-        rate_limiter = get_rate_limiter()
-        status = check_before_call("opencode")
-
-        if not status.is_available:
-            logger.warning(f"OpenCode not available: {status.error}")
-            return None
-
-        if status.wait_seconds > 0:
-            logger.info(
-                f"Waiting {status.wait_seconds:.1f}s for OpenCode rate limit..."
-            )
-            time.sleep(status.wait_seconds)
-
-        # Proactive rate limiting
-        elapsed = time.time() - self._last_call_time
-        if elapsed < self.MIN_CALL_INTERVAL:
-            time.sleep(self.MIN_CALL_INTERVAL - elapsed)
-
-        # Free models to try in order
-        free_models = ["glm-4.7-free", "minimax-m2.1-free"]
-
-        for model in free_models:
-            for attempt in range(max_retries):
-                try:
-                    self._last_call_time = time.time()
-                    logger.info(
-                        f"Trying OpenCode {model} (attempt {attempt + 1}/{max_retries})"
-                    )
-                    response = self.session.post(
-                        "https://opencode.ai/zen/v1/chat/completions",
-                        headers={
-                            "Authorization": f"Bearer {opencode_key}",
-                            "Content-Type": "application/json",
-                        },
-                        json={
-                            "model": model,
-                            "messages": [{"role": "user", "content": prompt}],
-                            "max_tokens": max_tokens,
-                            "temperature": 0.7,
-                        },
-                        timeout=60,
-                    )
-                    response.raise_for_status()
-
-                    # Update rate limiter from response headers
-                    rate_limiter.update_from_response_headers(
-                        "opencode", dict(response.headers)
-                    )
-                    rate_limiter._last_call_time["opencode"] = time.time()
-
-                    result = (
-                        response.json()
-                        .get("choices", [{}])[0]
-                        .get("message", {})
-                        .get("content")
-                    )
-                    if result:
-                        logger.info(f"OpenCode success with {model}")
-                        return result
-
-                except requests.exceptions.HTTPError as e:
-                    if response.status_code == 429:
-                        retry_after = response.headers.get("Retry-After", "10")
-                        try:
-                            wait_time = min(float(retry_after), self.MAX_RETRY_WAIT)
-                        except ValueError:
-                            wait_time = self.MAX_RETRY_WAIT
-                        logger.warning(
-                            f"OpenCode rate limited, waiting {wait_time}s (attempt {attempt + 1}/{max_retries})"
-                        )
-                        time.sleep(wait_time)
-                        continue
-                    logger.warning(f"OpenCode API error with {model}: {e}")
-                    break  # Try next model
-                except Exception as e:
-                    logger.warning(f"OpenCode API error with {model}: {e}")
-                    break  # Try next model
-
-        logger.warning("All OpenCode models failed")
-        return None
+        timing = [self._last_call_time]
+        result = call_openai_compatible(
+            OPENCODE_SPEC, opencode_key, prompt, max_tokens, max_retries,
+            self.session, self.MAX_RETRY_WAIT, self.MIN_CALL_INTERVAL, timing,
+        )
+        self._last_call_time = timing[0]
+        return result
 
     def _call_huggingface(
         self, prompt: str, max_tokens: int = 800, max_retries: int = 1
@@ -2155,91 +1942,13 @@ DATE: {datetime.now().strftime('%B %d, %Y')}"""
     ) -> Optional[str]:
         """Call Mistral AI API - high quality free tier models."""
         mistral_key = os.getenv("MISTRAL_API_KEY")
-        if not mistral_key:
-            return None
-
-        # Check rate limits before calling
-        rate_limiter = get_rate_limiter()
-        status = check_before_call("mistral")
-
-        if not status.is_available:
-            logger.warning(f"Mistral not available: {status.error}")
-            return None
-
-        if status.wait_seconds > 0:
-            logger.info(f"Waiting {status.wait_seconds:.1f}s for Mistral rate limit...")
-            time.sleep(status.wait_seconds)
-
-        # Proactive rate limiting
-        elapsed = time.time() - self._last_call_time
-        if elapsed < self.MIN_CALL_INTERVAL:
-            time.sleep(self.MIN_CALL_INTERVAL - elapsed)
-
-        # Mistral free tier models
-        models = [
-            "mistral-small-latest",
-            "open-mistral-7b",
-        ]
-
-        for model in models:
-            for attempt in range(max_retries):
-                try:
-                    self._last_call_time = time.time()
-                    logger.info(
-                        f"Trying Mistral {model} (attempt {attempt + 1}/{max_retries})"
-                    )
-                    response = self.session.post(
-                        "https://api.mistral.ai/v1/chat/completions",
-                        headers={
-                            "Authorization": f"Bearer {mistral_key}",
-                            "Content-Type": "application/json",
-                        },
-                        json={
-                            "model": model,
-                            "messages": [{"role": "user", "content": prompt}],
-                            "max_tokens": max_tokens,
-                            "temperature": 0.7,
-                        },
-                        timeout=60,
-                    )
-                    response.raise_for_status()
-
-                    # Update rate limiter from response headers
-                    rate_limiter.update_from_response_headers(
-                        "mistral", dict(response.headers)
-                    )
-                    rate_limiter._last_call_time["mistral"] = time.time()
-
-                    result = (
-                        response.json()
-                        .get("choices", [{}])[0]
-                        .get("message", {})
-                        .get("content")
-                    )
-                    if result:
-                        logger.info(f"Mistral success with {model}")
-                        return result
-
-                except requests.exceptions.HTTPError as e:
-                    if response.status_code == 429:
-                        retry_after = response.headers.get("Retry-After", "10")
-                        try:
-                            wait_time = min(float(retry_after), self.MAX_RETRY_WAIT)
-                        except ValueError:
-                            wait_time = self.MAX_RETRY_WAIT
-                        logger.warning(
-                            f"Mistral rate limited, waiting {wait_time}s (attempt {attempt + 1}/{max_retries})"
-                        )
-                        time.sleep(wait_time)
-                        continue
-                    logger.warning(f"Mistral API error with {model}: {e}")
-                    break  # Try next model
-                except Exception as e:
-                    logger.warning(f"Mistral API error with {model}: {e}")
-                    break  # Try next model
-
-        logger.warning("All Mistral models failed")
-        return None
+        timing = [self._last_call_time]
+        result = call_openai_compatible(
+            MISTRAL_SPEC, mistral_key, prompt, max_tokens, max_retries,
+            self.session, self.MAX_RETRY_WAIT, self.MIN_CALL_INTERVAL, timing,
+        )
+        self._last_call_time = timing[0]
+        return result
 
     def _repair_json(self, json_str: str) -> str:
         """Attempt to repair common JSON formatting issues from LLM output."""
