+++ scripts/main.py
@@ -50,13 +50,8 @@ from sitemap_generator import save_sitemap
 from editorial_generator import EditorialGenerator
 from fetch_media_of_day import MediaOfDayFetcher
 from cmmc_page_generator import generate_cmmc_page, filter_cmmc_trends
-from shared_components import (
-    build_header,
-    build_footer,
-    get_header_styles,
-    get_footer_styles,
-    get_theme_script,
-)
+from topic_page_generator import generate_all_topic_pages
+from media_page_generator import generate_media_page
 from image_utils import (
     validate_image_url,
     sanitize_image_url,
@@ -888,15 +883,11 @@ class Pipeline:
     def _step_generate_topic_pages(self) -> None:
         """Generate topic-specific sub-pages (/tech, /world, /science, etc.)."""
         logger.info("[9/16] Generating topic sub-pages...")
-
-        # Convert data to proper format
         trends_data = [
             asdict(t) if hasattr(t, "__dataclass_fields__") else t for t in self.trends
         ]
         design_data = (
-            asdict(self.design)
-            if hasattr(self.design, "__dataclass_fields__")
-            else self.design
+            asdict(self.design) if hasattr(self.design, "__dataclass_fields__") else self.design
         )
         images_data = [
             asdict(i) if hasattr(i, "__dataclass_fields__") else i for i in self.images
@@ -904,387 +895,12 @@ class Pipeline:
 
         self._apply_story_summaries(trends_data)
 
-        # Define topic categories and their filters
-        # Sources use prefix matching: 'tech_' matches 'tech_verge', 'tech_wired', etc.
-        # hero_keywords used to find relevant hero images for each topic
-        topic_configs = [
-            {
-                "slug": "tech",
-                "title": "Technology",
-                "description": "Latest technology news, startups, and developer trends",
-                "source_prefixes": [
-                    "hackernews",
-                    "lobsters",
-                    "tech_",
-                    "github_trending",
-                    "product_hunt",
-                    "devto",
-                    "slashdot",
-                    "ars_",
-                ],
-                "hero_keywords": [
-                    "technology",
-                    "computer",
-                    "code",
-                    "programming",
-                    "software",
-                    "digital",
-                    "tech",
-                    "innovation",
-                    "startup",
-                ],
-                "image_index": 0,  # Fallback index if no keyword match
-            },
-            {
-                "slug": "world",
-                "title": "World News",
-                "description": "Breaking news and current events from around the world",
-                "source_prefixes": ["news_", "wikipedia", "google_trends"],
-                "hero_keywords": [
-                    "world",
-                    "globe",
-                    "city",
-                    "cityscape",
-                    "urban",
-                    "international",
-                    "news",
-                    "global",
-                    "earth",
-                ],
-                "image_index": 1,
-            },
-            {
-                "slug": "science",
-                "title": "Science & Health",
-                "description": "Latest discoveries in science, technology, medicine, and space",
-                "source_prefixes": ["science_"],
-                "hero_keywords": [
-                    "science",
-                    "laboratory",
-                    "research",
-                    "space",
-                    "medical",
-                    "health",
-                    "biology",
-                    "chemistry",
-                    "physics",
-                ],
-                "image_index": 2,
-            },
-            {
-                "slug": "politics",
-                "title": "Politics & Policy",
-                "description": "Political news, policy analysis, and government updates",
-                "source_prefixes": ["politics_"],
-                "hero_keywords": [
-                    "politics",
-                    "government",
-                    "capitol",
-                    "democracy",
-                    "vote",
-                    "election",
-                    "law",
-                    "justice",
-                    "congress",
-                ],
-                "image_index": 3,
-            },
-            {
-                "slug": "finance",
-                "title": "Business & Finance",
-                "description": "Market news, business trends, and economic analysis",
-                "source_prefixes": ["finance_"],
-                "hero_keywords": [
-                    "finance",
-                    "business",
-                    "money",
-                    "stock",
-                    "market",
-                    "office",
-                    "corporate",
-                    "economy",
-                    "trading",
-                ],
-                "image_index": 4,
-            },
-            {
-                "slug": "business",
-                "title": "Business",
-                "description": "Latest business news, entrepreneurship, and corporate trends",
-                "source_prefixes": ["finance_", "business"],
-                "hero_keywords": [
-                    "business",
-                    "entrepreneur",
-                    "startup",
-                    "corporate",
-                    "office",
-                    "meeting",
-                    "professional",
-                    "commerce",
-                    "trade",
-                ],
-                "image_index": 5,
-            },
-            {
-                "slug": "sports",
-                "title": "Sports",
-                "description": "Latest sports news, scores, and athletic highlights",
-                "source_prefixes": ["sports_"],
-                "hero_keywords": [
-                    "sports",
-                    "athlete",
-                    "game",
-                    "stadium",
-                    "competition",
-                    "fitness",
-                    "team",
-                    "basketball",
-                    "football",
-                ],
-                "image_index": 6,
-            },
-        ]
-
-        def find_topic_image(
-            images: List[Record],
-            headline: str,
-            category_keywords: List[str],
-            fallback_index: int,
-            used_image_ids: Set[str],
-        ) -> Record:
-            """Find an image matching headline content, excluding already-used images.
-
-            Priority:
-            1. Match keywords from the actual headline (top story title)
-            2. Fall back to generic category keywords
-            3. Use fallback index if no matches (cycling through unused images)
-
-            Args:
-                images: List of available images
-                headline: The headline text to match
-                category_keywords: Fallback keywords for the category
-                fallback_index: Index for fallback selection
-                used_image_ids: Set of image IDs already used (will be modified)
-
-            Returns:
-                Best matching image dict, or empty dict if none available
-            """
-            if not images:
-                return {}
-
-            # Filter out already-used images to ensure each topic page gets a unique image
-            available_images = [
-                img for img in images if img.get("id") not in used_image_ids
-            ]
-            if not available_images:
-                # If all images used, reset and allow reuse (better than no image)
-                available_images = images
-
-            # Extract keywords from headline (similar to _find_relevant_hero_image)
-            stop_words = {
-                "the",
-                "a",
-                "an",
-                "is",
-                "are",
-                "was",
-                "were",
-                "be",
-                "been",
-                "being",
-                "have",
-                "has",
-                "had",
-                "do",
-                "does",
-                "did",
-                "will",
-                "would",
-                "could",
-                "should",
-                "may",
-                "might",
-                "must",
-                "shall",
-                "can",
-                "of",
-                "in",
-                "to",
-                "for",
-                "on",
-                "with",
-                "at",
-                "by",
-                "from",
-                "as",
-                "into",
-                "through",
-                "and",
-                "or",
-                "but",
-                "if",
-                "then",
-                "than",
-                "so",
-                "that",
-                "this",
-                "what",
-                "which",
-                "who",
-                "whom",
-                "how",
-                "when",
-                "where",
-                "why",
-                "says",
-                "said",
-                "new",
-                "first",
-                "after",
-                "year",
-                "years",
-                "now",
-                "today's",
-                "trends",
-                "trending",
-                "world",
-                "its",
-                "it",
-                "just",
-            }
-
-            headline_lower = headline.lower()
-            words = [w.strip(".,!?()[]{}\":;'") for w in headline_lower.split()]
-            headline_keywords = [w for w in words if len(w) > 2 and w not in stop_words]
-
-            # Score images - prioritize headline keywords over category keywords
-            best_image: Optional[Record] = None
-            best_score = 0.0
-
-            for img in available_images:
-                img_text = f"{img.get('query', '')} {img.get('description', '')} {img.get('alt', '')}".lower()
-
-                # Score based on headline keywords (weighted higher)
-                headline_score = sum(2 for kw in headline_keywords if kw in img_text)
-
-                # Add score for category keywords (weighted lower)
-                category_score = sum(1 for kw in category_keywords if kw in img_text)
-
-                total_score = float(headline_score + category_score)
-
-                # Prefer larger images
-                if img.get("width", 0) >= 1200:
-                    total_score += 0.5
-
-                if total_score > best_score:
-                    best_score = total_score
-                    best_image = img
-
-            # If found a match, use it
-            if best_image and best_score > 0:
-                if best_image.get("id"):
-                    used_image_ids.add(best_image["id"])
-                return best_image
-
-            # Otherwise use fallback index (cycling through available images)
-            idx = fallback_index % len(available_images)
-            selected = available_images[idx]
-            if selected.get("id"):
-                used_image_ids.add(selected["id"])
-            return selected
-
-        def matches_topic(source: str, prefixes: List[str]) -> bool:
-            """Check if a source matches any of the topic's prefixes."""
-            for prefix in prefixes:
-                if prefix.endswith("_"):
-                    if source.startswith(prefix):
-                        return True
-                else:
-                    if source == prefix:
-                        return True
-            return False
-
-        pages_created = 0
-        used_image_ids: Set[str] = set()
-
-        for config in topic_configs:
-            source_prefixes = config.get("source_prefixes", [])
-            if not isinstance(source_prefixes, list):
-                continue
-            source_prefixes = [str(prefix) for prefix in source_prefixes]
-            slug = str(config.get("slug", "")).strip()
-            if not slug:
-                continue
-
-            # Filter trends for this topic using prefix matching
-            topic_trends = [
-                t
-                for t in trends_data
-                if matches_topic(str(t.get("source", "")), source_prefixes)
-            ]
-
-            if len(topic_trends) < 3:
-                logger.info(
-                    f"  Skipping /{slug}/ - only {len(topic_trends)} stories"
-                )
-                continue
-
-            # Get the top story for hero image matching
-            top_story = topic_trends[0] if topic_trends else {}
-            top_story_title = top_story.get("title", "")
-
-            # Priority 1: Use article image from RSS feed if available
-            article_image_url = top_story.get("image_url")
-            if article_image_url:
-                # Create hero_image dict from article image
-                hero_image = {
-                    "url_large": article_image_url,
-                    "url_medium": article_image_url,
-                    "url_original": article_image_url,
-                    "photographer": "Article Image",
-                    "source": "article",
-                    "alt": top_story_title,
-                    "id": f"article_{hash(article_image_url) % 100000}",
-                }
-                logger.debug(
-                    f"  Using article image for {slug}: {article_image_url[:60]}..."
-                )
-            else:
-                hero_keywords_value = config.get("hero_keywords", [])
-                hero_keywords = (
-                    [str(keyword) for keyword in hero_keywords_value]
-                    if isinstance(hero_keywords_value, list)
-                    else []
-                )
-                image_index_value = config.get("image_index", 0)
-                image_index = (
-                    int(image_index_value)
-                    if isinstance(image_index_value, (int, float))
-                    else 0
-                )
-                # Priority 2: Fall back to stock photo search
-                hero_image = find_topic_image(
-                    images_data,
-                    top_story_title,
-                    hero_keywords,
-                    image_index,
-                    used_image_ids,
-                )
-
-            # Create topic directory
-            topic_dir = self.public_dir / slug
-            topic_dir.mkdir(parents=True, exist_ok=True)
-
-            # Build topic page HTML
-            html = self._build_topic_page(config, topic_trends, design_data, hero_image)
-
-            # Save
-            (topic_dir / "index.html").write_text(html, encoding="utf-8")
-            pages_created += 1
-            logger.info(
-                f"  Created /{slug}/ with {len(topic_trends)} stories"
-            )
-
+        pages_created = generate_all_topic_pages(
+            public_dir=self.public_dir,
+            trends_data=trends_data,
+            images_data=images_data,
+            design_data=design_data,
+        )
         logger.info(f"Generated {pages_created} topic sub-pages")
         self.metrics.set_counter("topic_pages_created", pages_created)
 
@@ -1333,748 +949,6 @@ class Pipeline:
             logger.warning("  Failed to generate CMMC page")
             self.metrics.set_counter("cmmc_page_generated", 0)
 
-    def _build_topic_page(
-        self,
-        config: Record,
-        trends: List[Record],
-        design: Record,
-        hero_image: Record,
-    ) -> str:
-        """Build HTML for a topic sub-page with shared header/footer."""
-        from datetime import datetime
-        import html as html_module
-
-        colors = {
-            "bg": design.get("color_bg", "#0a0a0a"),
-            "card_bg": design.get("color_card_bg", "#18181b"),
-            "text": design.get("color_text", "#ffffff"),
-            "muted": design.get("color_muted", "#a1a1aa"),
-            "border": design.get("color_border", "#27272a"),
-            "accent": design.get("color_accent", "#6366f1"),
-            "accent_secondary": design.get("color_accent_secondary", "#8b5cf6"),
-        }
-        font_primary = design.get("font_primary", "Space Grotesk")
-        font_secondary = design.get("font_secondary", "Inter")
-        radius = design.get("card_radius", "1rem")
-        card_padding = design.get("card_padding", "1.5rem")
-        transition = design.get("transition_speed", "200ms")
-        base_mode = "dark-mode" if design.get("is_dark_mode", True) else "light-mode"
-
-        # Get date info
-        now = datetime.now()
-        date_str = now.strftime("%B %d, %Y")
-        date_iso = now.isoformat()
-
-        # Get hero image URL and alt text (topic-specific image passed in)
-        hero_image_url = ""
-        hero_image_alt = ""
-        if hero_image:
-            hero_image_url = str(
-                hero_image.get(
-                    "url_large",
-                    hero_image.get("url_medium", hero_image.get("url", "")),
-                )
-                or ""
-            )
-            hero_image_alt = str(
-                hero_image.get(
-                    "alt",
-                    hero_image.get(
-                        "description", f"{str(config.get('title', 'Topic'))} hero image"
-                    ),
-                )
-                or ""
-            )
-
-        # Get featured story info (handle None values safely)
-        featured_story = trends[0] if trends else {}
-        featured_title = html_module.escape((featured_story.get("title") or "")[:100])
-        featured_url = html_module.escape(featured_story.get("url") or "#")
-        featured_source_raw = featured_story.get("source_label")
-        if not featured_source_raw:
-            featured_source_raw = (featured_story.get("source") or "").replace(
-                "_", " "
-            ).title()
-        featured_source = html_module.escape(featured_source_raw)
-        featured_desc = html_module.escape(
-            (featured_story.get("summary") or featured_story.get("description") or "")[
-                :200
-            ]
-        )
-
-        # Placeholder image URL (gradient fallback from homepage)
-        placeholder_url = "/assets/nano-banana.png"
-
-        # Build story cards with enhanced design (skip first since it's in hero)
-        cards: List[str] = []
-        for i, t in enumerate(trends[1:20]):  # Start from index 1, skip featured
-            title = html_module.escape((t.get("title") or "")[:100])
-            url = html_module.escape(t.get("url") or "#")
-            source_raw = t.get("source_label")
-            if not source_raw:
-                source_raw = (t.get("source") or "").replace("_", " ").title()
-            source = html_module.escape(source_raw)
-            raw_image_url = t.get("image_url") or ""
-
-            # Validate and sanitize the image URL for reliability
-            is_valid, validated_url = validate_image_url(raw_image_url)
-
-            # Always show an image - use placeholder if no valid image available
-            if is_valid and validated_url:
-                img_src = html_module.escape(validated_url)
-                img_class = "story-image"
-                img_alt = title
-                # Add data attribute for quality score (helps with debugging)
-                img_quality = get_image_quality_score(validated_url)
-                img_data_attrs = f'data-quality="{img_quality}"'
-            else:
-                img_src = placeholder_url
-                img_class = "story-image placeholder"
-                img_alt = f"{source} story placeholder"
-                img_data_attrs = 'data-is-placeholder="true"'
-
-            cards.append(
-                f"""
-            <article class="story-card">
-                <div class="story-wrapper">
-                    <figure class="story-media">
-                        <img src="{img_src}"
-                             alt="{img_alt}"
-                             class="{img_class}"
-                             loading="lazy"
-                             referrerpolicy="no-referrer"
-                             width="640"
-                             height="360"
-                             {img_data_attrs}
-                             onerror="this.onerror=null;this.src='{placeholder_url}';this.classList.add('placeholder');">
-                    </figure>
-                    <div class="story-content">
-                        <span class="source-badge">{source}</span>
-                        <h3 class="story-title">
-                            <a href="{url}" target="_blank" rel="noopener">{title}</a>
-                        </h3>
-                    </div>
-                </div>
-            </article>"""
-            )
-
-        return f"""<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>{config['title']} | DailyTrending.info</title>
-    <meta name="description" content="{config['description']}">
-    <meta name="author" content="DailyTrending.info">
-    <meta name="robots" content="index, follow, max-image-preview:large, max-snippet:-1, max-video-preview:-1">
-    <link rel="canonical" href="https://dailytrending.info/{config['slug']}/">
-    <link rel="icon" type="image/svg+xml" href="/favicon.svg">
-
-    <meta property="og:title" content="{config['title']} | DailyTrending.info">
-    <meta property="og:description" content="{config['description']}">
-    <meta property="og:type" content="website">
-    <meta property="og:url" content="https://dailytrending.info/{config['slug']}/">
-    <meta property="og:image" content="https://dailytrending.info/og-image.png">
-    <meta property="og:site_name" content="DailyTrending.info">
-
-    <meta name="twitter:card" content="summary_large_image">
-    <meta name="twitter:site" content="@bradshannon">
-    <meta name="twitter:title" content="{config['title']} | DailyTrending.info">
-    <meta name="twitter:description" content="{config['description']}">
-
-    <!-- Google AdSense -->
-    <script async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-2196222970720414"
-         crossorigin="anonymous"></script>
-
-    <script type="application/ld+json">
-    {{
-        "@context": "https://schema.org",
-        "@graph": [
-            {{
-                "@type": "BreadcrumbList",
-                "itemListElement": [
-                    {{
-                        "@type": "ListItem",
-                        "position": 1,
-                        "name": "Home",
-                        "item": "https://dailytrending.info/"
-                    }},
-                    {{
-                        "@type": "ListItem",
-                        "position": 2,
-                        "name": "{config['title']}",
-                        "item": "https://dailytrending.info/{config['slug']}/"
-                    }}
-                ]
-            }},
-            {{
-                "@type": "CollectionPage",
-                "name": "{config['title']} Trends",
-                "description": "{config['description']}",
-                "url": "https://dailytrending.info/{config['slug']}/",
-                "isPartOf": {{"@id": "https://dailytrending.info"}},
-                "dateModified": "{date_iso}",
-                "numberOfItems": {len(trends)},
-                "publisher": {{
-                    "@type": "Organization",
-                    "name": "DailyTrending.info",
-                    "url": "https://dailytrending.info/"
-                }}
-            }}
-        ]
-    }}
-    </script>
-
-    <link href="https://fonts.googleapis.com/css2?family={font_primary.replace(' ', '+')}:wght@400;500;600;700;800&family={font_secondary.replace(' ', '+')}:wght@400;500;600;700&display=swap" rel="stylesheet">
-    <style>
-        :root {{
-            --color-bg: {colors['bg']};
-            --color-card-bg: {colors['card_bg']};
-            --color-text: {colors['text']};
-            --color-muted: {colors['muted']};
-            --color-border: {colors['border']};
-            --color-accent: {colors['accent']};
-            --color-accent-secondary: {colors['accent_secondary']};
-            --radius: {radius};
-            --card-padding: {card_padding};
-            --transition: {transition} ease;
-            --font-primary: '{font_primary}', system-ui, sans-serif;
-            --font-secondary: '{font_secondary}', system-ui, sans-serif;
-        }}
-
-        * {{ margin: 0; padding: 0; box-sizing: border-box; }}
-
-        body {{
-            font-family: var(--font-secondary);
-            background: var(--color-bg);
-            color: var(--color-text);
-            line-height: 1.6;
-            min-height: 100vh;
-        }}
-
-        body.light-mode {{
-            --color-bg: #ffffff;
-            --color-card-bg: #f8fafc;
-            --color-text: #1a1a2e;
-            --color-muted: #64748b;
-            --color-border: #e2e8f0;
-            background: var(--color-bg);
-        }}
-
-        body.dark-mode {{
-            --color-bg: #0a0a0a;
-            --color-card-bg: #18181b;
-            --color-text: #ffffff;
-            --color-muted: #a1a1aa;
-            --color-border: #27272a;
-            background: var(--color-bg);
-        }}
-
-        /* Density settings */
-        body.density-compact {{
-            --section-gap: 1.5rem;
-            --card-gap: 0.75rem;
-            --card-padding: 0.75rem;
-        }}
-        body.density-comfortable {{
-            --section-gap: 2.5rem;
-            --card-gap: 1.25rem;
-            --card-padding: 1.25rem;
-        }}
-        body.density-spacious {{
-            --section-gap: 4rem;
-            --card-gap: 2rem;
-            --card-padding: 1.75rem;
-        }}
-
-        /* View list mode */
-        body.view-list .stories-grid,
-        body.view-list .trend-grid {{
-            display: flex;
-            flex-direction: column;
-            gap: 0;
-        }}
-        body.view-list .story-card,
-        body.view-list .trend-card {{
-            background: transparent;
-            border: none;
-            border-bottom: 1px solid var(--color-border);
-            border-radius: 0;
-            padding: 0.5rem 0;
-            padding-left: 1.5rem;
-        }}
-        body.view-list .story-card img,
-        body.view-list .trend-card img,
-        body.view-list .card-image {{
-            display: none;
-        }}
-
-        {get_header_styles()}
-
-        /* Hero Header with Featured Story */
-        .topic-hero {{
-            position: relative;
-            min-height: 500px;
-            display: flex;
-            align-items: flex-end;
-            overflow: hidden;
-            border-bottom: 1px solid var(--color-border);
-        }}
-
-        .hero-image {{
-            position: absolute;
-            inset: 0;
-            background-size: contain;
-            background-position: center center;
-            background-repeat: no-repeat;
-            background-color: var(--color-bg);
-            z-index: 0;
-        }}
-
-        /* Darkened scaled version behind for full coverage */
-        .hero-image::before {{
-            content: '';
-            position: absolute;
-            inset: -20px;
-            background: inherit;
-            background-size: cover;
-            filter: brightness(0.5);
-            z-index: -1;
-        }}
-
-        .hero-image::after {{
-            content: '';
-            position: absolute;
-            inset: 0;
-            background: linear-gradient(to top, rgba(0,0,0,0.9) 0%, rgba(0,0,0,0.4) 50%, rgba(0,0,0,0.2) 100%);
-        }}
-
-        .hero-content {{
-            position: relative;
-            z-index: 1;
-            width: 100%;
-            max-width: 1200px;
-            margin: 0 auto;
-            padding: 3rem 2rem;
-        }}
-
-        .topic-label {{
-            display: inline-flex;
-            align-items: center;
-            gap: 0.5rem;
-            background: var(--color-accent);
-            color: #000;
-            padding: 0.4rem 1rem;
-            border-radius: 999px;
-            font-size: 0.75rem;
-            font-weight: 700;
-            text-transform: uppercase;
-            letter-spacing: 0.1em;
-            margin-bottom: 1rem;
-        }}
-
-        .hero-title {{
-            font-family: var(--font-primary);
-            font-size: clamp(1.75rem, 4vw, 2.75rem);
-            font-weight: 700;
-            line-height: 1.2;
-            margin-bottom: 1rem;
-            max-width: 800px;
-        }}
-
-        .hero-title a {{
-            color: var(--color-text);
-            text-decoration: none;
-            transition: color var(--transition);
-        }}
-
-        .hero-title a:hover {{
-            color: var(--color-accent);
-        }}
-
-        .hero-desc {{
-            font-size: 1.1rem;
-            color: var(--color-muted);
-            max-width: 600px;
-            margin-bottom: 1.5rem;
-            line-height: 1.6;
-        }}
-
-        .hero-meta {{
-            display: flex;
-            align-items: center;
-            gap: 1.5rem;
-            flex-wrap: wrap;
-        }}
-
-        .hero-source {{
-            display: inline-flex;
-            align-items: center;
-            gap: 0.5rem;
-            color: var(--color-accent);
-            font-weight: 600;
-            font-size: 0.9rem;
-        }}
-
-        .hero-stats {{
-            display: flex;
-            gap: 1.5rem;
-            font-size: 0.9rem;
-            color: var(--color-muted);
-        }}
-
-        .hero-stats span {{
-            display: flex;
-            align-items: center;
-            gap: 0.5rem;
-        }}
-
-        .hero-cta {{
-            display: inline-flex;
-            align-items: center;
-            gap: 0.5rem;
-            padding: 0.75rem 1.5rem;
-            background: var(--color-accent);
-            color: #000;
-            font-weight: 600;
-            border-radius: var(--radius);
-            text-decoration: none;
-            transition: transform var(--transition), box-shadow var(--transition);
-        }}
-
-        .hero-cta:hover {{
-            transform: translateY(-2px);
-            box-shadow: 0 10px 25px rgba(0,0,0,0.3);
-        }}
-
-        /* Main Content */
-        .main-content {{
-            max-width: 1200px;
-            margin: 0 auto;
-            padding: 3rem 2rem;
-        }}
-
-        .stories-grid {{
-            display: grid;
-            grid-template-columns: repeat(4, 1fr);
-            gap: 1.25rem;
-        }}
-
-        .story-card {{
-            background: var(--color-card-bg);
-            border: 1px solid var(--color-border);
-            border-radius: var(--radius);
-            overflow: hidden;
-            display: flex;
-            flex-direction: column;
-            transition: transform var(--transition), border-color var(--transition), box-shadow var(--transition);
-        }}
-
-        .story-card:hover {{
-            transform: translateY(-4px);
-            border-color: var(--color-accent);
-            box-shadow: 0 20px 40px rgba(0,0,0,0.3);
-        }}
-
-        .story-wrapper {{
-            display: flex;
-            flex-direction: column;
-            width: 100%;
-            height: 100%;
-        }}
-
-        .story-media {{
-            width: 100%;
-            flex-shrink: 0;
-            border-radius: 0;
-            overflow: hidden;
-            position: relative;
-            background: color-mix(in srgb, rgba(12, 16, 24, 0.95), rgba(34, 45, 63, 0.9));
-            background-image: radial-gradient(circle at 30% 25%, rgba(255, 255, 255, 0.18), transparent 40%),
-                              radial-gradient(circle at 70% 80%, rgba(255, 255, 255, 0.08), transparent 55%);
-            min-height: 180px;
-        }}
-
-        .story-media::before {{
-            content: '';
-            position: absolute;
-            inset: 0;
-            background: linear-gradient(to bottom, rgba(0, 0, 0, 0.02), rgba(0, 0, 0, 0.25));
-            pointer-events: none;
-        }}
-
-        .story-image {{
-            width: 100%;
-            height: 180px;
-            min-height: 180px;
-            object-fit: cover;
-            object-position: center;
-            background-color: var(--color-border);
-            transition: opacity 0.3s ease;
-        }}
-
-        /* Loading state - shimmer effect */
-        .story-image:not([loaded]):not(.placeholder) {{
-            opacity: 0;
-        }}
-
-        .story-media:not(.image-loaded):not(.image-fallback)::after {{
-            content: '';
-            position: absolute;
-            inset: 0;
-            background: linear-gradient(90deg,
-                transparent 0%,
-                rgba(255, 255, 255, 0.08) 50%,
-                transparent 100%);
-            animation: shimmer 1.5s infinite;
-        }}
-
-        @keyframes shimmer {{
-            0% {{ transform: translateX(-100%); }}
-            100% {{ transform: translateX(100%); }}
-        }}
-
-        /* Loaded state */
-        .story-image[loaded] {{
-            opacity: 1;
-        }}
-
-        .story-image.placeholder {{
-            opacity: 0.85;
-            filter: grayscale(0.1);
-        }}
-
-        /* Fallback indicator (subtle) */
-        .image-fallback .story-image {{
-            opacity: 0.8;
-        }}
-
-        .story-content {{
-            padding: 1rem;
-            flex: 1;
-            display: flex;
-            flex-direction: column;
-        }}
-
-        .source-badge {{
-            display: inline-block;
-            font-size: 0.7rem;
-            font-weight: 600;
-            text-transform: uppercase;
-            letter-spacing: 0.05em;
-            color: var(--color-accent);
-            margin-bottom: 0.4rem;
-        }}
-
-        .story-title {{
-            font-size: 0.95rem;
-            font-weight: 600;
-            margin-bottom: 0.5rem;
-            line-height: 1.4;
-            display: -webkit-box;
-            -webkit-line-clamp: 2;
-            -webkit-box-orient: vertical;
-            overflow: hidden;
-        }}
-
-        .story-title a {{
-            color: var(--color-text);
-            text-decoration: none;
-            background: linear-gradient(to right, var(--color-accent), var(--color-accent)) 0 100% / 0 2px no-repeat;
-            transition: background-size 0.3s;
-        }}
-
-        .story-title a:hover {{
-            background-size: 100% 2px;
-        }}
-
-        .story-desc {{
-            color: var(--color-muted);
-            font-size: 0.8rem;
-            line-height: 1.5;
-            margin-bottom: 0.75rem;
-            display: -webkit-box;
-            -webkit-line-clamp: 2;
-            -webkit-box-orient: vertical;
-            overflow: hidden;
-        }}
-
-        {get_footer_styles()}
-
-        /* Responsive - Mobile (1 column) */
-        @media (max-width: 640px) {{
-            .stories-grid {{
-                grid-template-columns: 1fr;
-            }}
-
-            .topic-hero {{
-                min-height: 350px;
-            }}
-
-            .hero-content {{
-                padding: 2rem 1rem;
-            }}
-
-            .hero-title {{
-                font-size: 1.5rem;
-            }}
-
-            .hero-desc {{
-                font-size: 1rem;
-            }}
-
-            .hero-meta {{
-                flex-direction: column;
-                align-items: flex-start;
-                gap: 1rem;
-            }}
-
-            .main-content {{
-                padding: 2rem 1rem;
-            }}
-
-            .story-card.featured {{
-                grid-column: 1;
-            }}
-        }}
-
-        /* Responsive - Tablet (2 columns) */
-        @media (min-width: 641px) and (max-width: 1024px) {{
-            .stories-grid {{
-                grid-template-columns: repeat(2, 1fr);
-            }}
-
-            .topic-hero {{
-                min-height: 400px;
-            }}
-        }}
-
-        /* Responsive - Small Desktop (3 columns) */
-        @media (min-width: 1025px) and (max-width: 1280px) {{
-            .stories-grid {{
-                grid-template-columns: repeat(3, 1fr);
-            }}
-        }}
-    </style>
-</head>
-<body class="{base_mode}">
-    {build_header(config['slug'], date_str)}
-
-    <header class="topic-hero">
-        <div class="hero-image" style="background-image: url('{hero_image_url}');" role="img" aria-label="{hero_image_alt}"></div>
-        <div class="hero-content">
-            <span class="topic-label">{config['title']}</span>
-            <h1 class="hero-title"><a href="{featured_url}" target="_blank" rel="noopener">{featured_title}</a></h1>
-            {f'<p class="hero-desc">{featured_desc}</p>' if featured_desc else ''}
-            <div class="hero-meta">
-                <span class="hero-source">
-                    <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><path d="M21 10c0 7-9 13-9 13s-9-6-9-13a9 9 0 0 1 18 0z"/><circle cx="12" cy="10" r="3"/></svg>
-                    {featured_source}
-                </span>
-                <div class="hero-stats">
-                    <span>
-                        <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><path d="M12 20V10M18 20V4M6 20v-4"/></svg>
-                        {len(trends)} stories
-                    </span>
-                    <span>
-                        <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><circle cx="12" cy="12" r="10"/><polyline points="12 6 12 12 16 14"/></svg>
-                        {date_str}
-                    </span>
-                </div>
-            </div>
-        </div>
-    </header>
-
-    <main class="main-content">
-        <div class="stories-grid">
-            {''.join(cards)}
-        </div>
-    </main>
-
-    {build_footer(date_str)}
-
-    <script>
-        (function () {{
-            const placeholderImage = "{placeholder_url}";
-            const LOAD_TIMEOUT_MS = 8000; // 8 second timeout for slow images
-
-            const markBroken = (img, reason = 'error') => {{
-                if (!placeholderImage) return;
-                if (img.dataset.fallbackApplied) return;
-                img.dataset.fallbackApplied = 'true';
-                img.dataset.fallbackReason = reason;
-                img.src = placeholderImage;
-                img.classList.add('placeholder');
-                img.title = 'Image unavailable';
-                // Add visual indicator class for styling
-                img.parentElement?.classList.add('image-fallback');
-            }};
-
-            const bindErrors = () => {{
-                document.querySelectorAll('.story-image').forEach((img) => {{
-                    if (img.dataset.boundError) return;
-                    img.dataset.boundError = 'true';
-
-                    // Skip if already a placeholder
-                    if (img.dataset.isPlaceholder === 'true') {{
-                        img.setAttribute('loaded', '');
-                        return;
-                    }}
-
-                    // Set up timeout for slow-loading images
-                    let loadTimeout;
-                    const startTimeout = () => {{
-                        loadTimeout = setTimeout(() => {{
-                            if (!img.complete || img.naturalWidth === 0) {{
-                                markBroken(img, 'timeout');
-                            }}
-                        }}, LOAD_TIMEOUT_MS);
-                    }};
-
-                    // Handle successful image load
-                    img.addEventListener('load', () => {{
-                        clearTimeout(loadTimeout);
-                        if (img.naturalWidth > 0) {{
-                            img.setAttribute('loaded', '');
-                            img.parentElement?.classList.add('image-loaded');
-                        }} else {{
-                            markBroken(img, 'zero-size');
-                        }}
-                    }});
-
-                    // Handle image error
-                    img.addEventListener('error', () => {{
-                        clearTimeout(loadTimeout);
-                        markBroken(img, 'error');
-                    }});
-
-                    // Handle already-loaded images
-                    if (img.complete) {{
-                        if (img.naturalWidth === 0) {{
-                            markBroken(img, 'already-broken');
-                        }} else {{
-                            img.setAttribute('loaded', '');
-                            img.parentElement?.classList.add('image-loaded');
-                        }}
-                    }} else {{
-                        // Start timeout for images still loading
-                        startTimeout();
-                    }}
-                }});
-            }};
-
-            if (document.readyState === 'loading') {{
-                document.addEventListener('DOMContentLoaded', bindErrors);
-            }} else {{
-                bindErrors();
-            }}
-        }})();
-    </script>
-
-    {get_theme_script()}
-</body>
-</html>"""
-
     def _step_fetch_media_of_day(self) -> None:
         """Fetch image and video of the day from curated sources."""
         logger.info("[10/16] Fetching Media of the Day...")
@@ -2116,683 +990,13 @@ class Pipeline:
             self.metrics.set_counter("media_page_generated", 0)
             return
 
-        # Get design data for styling
         design_data = (
-            asdict(self.design)
-            if hasattr(self.design, "__dataclass_fields__")
-            else self.design
+            asdict(self.design) if hasattr(self.design, "__dataclass_fields__") else self.design
         )
-
-        # Create media directory
-        media_dir = self.public_dir / "media"
-        media_dir.mkdir(parents=True, exist_ok=True)
-
-        # Build media page HTML
-        html = self._build_media_page(self.media_data, design_data)
-
-        # Save
-        (media_dir / "index.html").write_text(html, encoding="utf-8")
-        logger.info(f"Media page saved to {media_dir / 'index.html'}")
-        self.metrics.set_counter("media_page_generated", 1)
-
-    def _build_media_page(self, media_data: Record, design: Record) -> str:
-        """Build HTML for the Media of the Day page."""
-        from datetime import datetime
-        import html as html_module
-
-        now = datetime.now()
-        date_str = now.strftime("%B %d, %Y")
-        date_iso = now.isoformat()
-
-        colors = {
-            "bg": design.get("color_bg", "#0a0a0a"),
-            "card_bg": design.get("color_card_bg", "#18181b"),
-            "text": design.get("color_text", "#ffffff"),
-            "muted": design.get("color_muted", "#a1a1aa"),
-            "border": design.get("color_border", "#27272a"),
-            "accent": design.get("color_accent", "#6366f1"),
-            "accent_secondary": design.get("color_accent_secondary", "#8b5cf6"),
-        }
-        font_primary = design.get("font_primary", "Space Grotesk")
-        font_secondary = design.get("font_secondary", "Inter")
-        radius = design.get("card_radius", "1rem")
-        transition = design.get("transition_speed", "200ms")
-        base_mode = "dark-mode" if design.get("is_dark_mode", True) else "light-mode"
-
-        # Helper to ensure value is a string before escaping
-        def safe_str(val: Any, default: str = "") -> str:
-            if val is None:
-                return default
-            if isinstance(val, list):
-                return val[0] if val else default
-            return str(val)
-
-        # Get image data
-        image = media_data.get("image_of_day") or {}
-        image_title = html_module.escape(
-            safe_str(image.get("title"), "Image of the Day")
-        )
-        image_url = html_module.escape(safe_str(image.get("url")))
-        image_hd_url = html_module.escape(safe_str(image.get("url_hd")))
-        image_explanation = html_module.escape(safe_str(image.get("explanation")))
-        image_source = safe_str(image.get("source"))
-        image_source_url = html_module.escape(safe_str(image.get("source_url")))
-        image_copyright = html_module.escape(safe_str(image.get("copyright")))
-        image_date = safe_str(image.get("date"))
-
-        # Get video data
-        video = media_data.get("video_of_day") or {}
-        video_title = html_module.escape(
-            safe_str(video.get("title"), "Video of the Day")
-        )
-        video_description = html_module.escape(safe_str(video.get("description")))
-        video_embed_url = html_module.escape(safe_str(video.get("embed_url")))
-        video_url = html_module.escape(safe_str(video.get("video_url")))
-        video_thumbnail = html_module.escape(safe_str(video.get("thumbnail_url")))
-        video_author = html_module.escape(safe_str(video.get("author")))
-        video_author_url = html_module.escape(safe_str(video.get("author_url")))
-        video_duration = safe_str(video.get("duration"))
-
-        # Source display names
-        source_names = {
-            "nasa_apod": "NASA Astronomy Picture of the Day",
-            "bing": "Bing Image of the Day",
-            "vimeo_staff_picks": "Vimeo Staff Picks",
-        }
-
-        image_source_name = source_names.get(image_source, image_source)
-        video_source_name = source_names.get(
-            video.get("source", ""), "Vimeo Staff Picks"
-        )
-
-        # Build conditional HTML parts to avoid nested f-string issues
-        copyright_html = ""
-        if image_copyright:
-            copyright_html = f"""<span class="meta-item">
-                            <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-                                <circle cx="12" cy="12" r="10"/>
-                                <path d="M14.31 8l5.74 9.94M9.69 8h11.48M7.38 12l5.74-9.94M9.69 16L3.95 6.06M14.31 16H2.83M16.62 12l-5.74 9.94"/>
-                            </svg>
-                            © {image_copyright}
-                        </span>"""
-
-        hd_link_html = ""
-        if image_hd_url:
-            hd_link_html = f"""<a href="{image_hd_url}" target="_blank" rel="noopener" class="action-btn secondary">
-                            <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-                                <path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4"/>
-                                <polyline points="7 10 12 15 17 10"/>
-                                <line x1="12" y1="15" x2="12" y2="3"/>
-                            </svg>
-                            HD Version
-                        </a>"""
-
-        author_html = ""
-        if video_author:
-            author_initial = video_author[0].upper() if video_author else "V"
-            author_html = f"""<div class="author-info">
-                        <div class="author-avatar">{author_initial}</div>
-                        <div>
-                            <span class="author-name">{video_author}</span>
-                        </div>
-                    </div>"""
-
-        # Build image section HTML
-        explanation_truncated = (
-            image_explanation[:800] + "..."
-            if len(image_explanation) > 800
-            else image_explanation
-        )
-        image_source_short = (
-            image_source_name.split()[0] if image_source_name else "Source"
-        )
-        if image:
-            image_section = f"""<section class="media-section">
-            <div class="section-header">
-                <h2 class="section-title">
-                    <span class="section-icon">
-                        <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="#000" stroke-width="2">
-                            <rect x="3" y="3" width="18" height="18" rx="2" ry="2"/>
-                            <circle cx="8.5" cy="8.5" r="1.5"/>
-                            <polyline points="21 15 16 10 5 21"/>
-                        </svg>
-                    </span>
-                    Image of the Day
-                </h2>
-                <a href="{image_source_url}" target="_blank" rel="noopener" class="source-link">
-                    {image_source_name}
-                    <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-                        <path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"/>
-                        <polyline points="15 3 21 3 21 9"/>
-                        <line x1="10" y1="14" x2="21" y2="3"/>
-                    </svg>
-                </a>
-            </div>
-            <div class="image-container">
-                <img src="{image_url}" alt="{image_title}" class="featured-image" loading="lazy">
-                <div class="image-info">
-                    <h3 class="media-title">{image_title}</h3>
-                    <p class="media-description">{explanation_truncated}</p>
-                    <div class="media-meta">
-                        <span class="meta-item">
-                            <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-                                <rect x="3" y="4" width="18" height="18" rx="2" ry="2"/>
-                                <line x1="16" y1="2" x2="16" y2="6"/>
-                                <line x1="8" y1="2" x2="8" y2="6"/>
-                                <line x1="3" y1="10" x2="21" y2="10"/>
-                            </svg>
-                            {image_date}
-                        </span>
-                        {copyright_html}
-                    </div>
-                    <div class="image-actions">
-                        <a href="{image_source_url}" target="_blank" rel="noopener" class="action-btn">
-                            <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-                                <path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"/>
-                                <polyline points="15 3 21 3 21 9"/>
-                                <line x1="10" y1="14" x2="21" y2="3"/>
-                            </svg>
-                            View on {image_source_short}
-                        </a>
-                        {hd_link_html}
-                    </div>
-                </div>
-            </div>
-        </section>"""
-        else:
-            image_section = '<p style="color: var(--color-muted); text-align: center; padding: 2rem;">Image of the Day is temporarily unavailable.</p>'
-
-        # Build video section HTML
-        description_truncated = (
-            video_description[:500] + "..."
-            if len(video_description) > 500
-            else video_description
-        )
-        if video:
-            video_section = f"""<section class="media-section">
-            <div class="section-header">
-                <h2 class="section-title">
-                    <span class="section-icon">
-                        <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="#000" stroke-width="2">
-                            <polygon points="5 3 19 12 5 21 5 3"/>
-                        </svg>
-                    </span>
-                    Video of the Day
-                </h2>
-                <a href="https://vimeo.com/channels/staffpicks" target="_blank" rel="noopener" class="source-link">
-                    {video_source_name}
-                    <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-                        <path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"/>
-                        <polyline points="15 3 21 3 21 9"/>
-                        <line x1="10" y1="14" x2="21" y2="3"/>
-                    </svg>
-                </a>
-            </div>
-            <div class="video-container">
-                <div class="video-embed">
-                    <iframe src="{video_embed_url}?title=0&byline=0&portrait=0" allow="autoplay; fullscreen; picture-in-picture" allowfullscreen></iframe>
-                </div>
-                <div class="video-info">
-                    <h3 class="media-title">{video_title}</h3>
-                    <p class="media-description">{description_truncated}</p>
-                    {author_html}
-                    <div class="image-actions">
-                        <a href="{video_url}" target="_blank" rel="noopener" class="action-btn">
-                            <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-                                <path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"/>
-                                <polyline points="15 3 21 3 21 9"/>
-                                <line x1="10" y1="14" x2="21" y2="3"/>
-                            </svg>
-                            Watch on Vimeo
-                        </a>
-                    </div>
-                </div>
-            </div>
-        </section>"""
-        else:
-            video_section = '<p style="color: var(--color-muted); text-align: center; padding: 2rem;">Video of the Day is temporarily unavailable.</p>'
-
-        return f"""<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Media of the Day | DailyTrending.info</title>
-    <meta name="description" content="Daily curated image and video content - featuring NASA's Astronomy Picture of the Day and Vimeo Staff Picks.">
-    <link rel="canonical" href="https://dailytrending.info/media/">
-    <link rel="icon" type="image/svg+xml" href="/favicon.svg">
-
-    <meta property="og:title" content="Media of the Day | DailyTrending.info">
-    <meta property="og:description" content="Daily curated image and video content">
-    <meta property="og:type" content="website">
-    <meta property="og:url" content="https://dailytrending.info/media/">
-    {f'<meta property="og:image" content="{image_url}">' if image_url else ''}
-
-    <meta name="twitter:card" content="summary_large_image">
-    <meta name="twitter:title" content="Media of the Day | DailyTrending.info">
-    <meta name="twitter:description" content="Daily curated image and video content">
-
-    <!-- Google AdSense -->
-    <script async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-2196222970720414"
-         crossorigin="anonymous"></script>
-
-    <link href="https://fonts.googleapis.com/css2?family={font_primary.replace(' ', '+')}:wght@400;500;600;700;800&family={font_secondary.replace(' ', '+')}:wght@400;500;600;700&display=swap" rel="stylesheet">
-    <style>
-        :root {{
-            --color-bg: {colors['bg']};
-            --color-card-bg: {colors['card_bg']};
-            --color-text: {colors['text']};
-            --color-muted: {colors['muted']};
-            --color-border: {colors['border']};
-            --color-accent: {colors['accent']};
-            --color-accent-secondary: {colors['accent_secondary']};
-            --radius: {radius};
-            --transition: {transition} ease;
-            --font-primary: '{font_primary}', system-ui, sans-serif;
-            --font-secondary: '{font_secondary}', system-ui, sans-serif;
-        }}
-
-        * {{ margin: 0; padding: 0; box-sizing: border-box; }}
-
-        body {{
-            font-family: var(--font-secondary);
-            background: var(--color-bg);
-            color: var(--color-text);
-            line-height: 1.6;
-            min-height: 100vh;
-        }}
-
-        body.light-mode {{
-            --color-bg: #ffffff;
-            --color-card-bg: #f8fafc;
-            --color-text: #1a1a2e;
-            --color-muted: #64748b;
-            --color-border: #e2e8f0;
-            background: var(--color-bg);
-        }}
-
-        body.dark-mode {{
-            --color-bg: #0a0a0a;
-            --color-card-bg: #18181b;
-            --color-text: #ffffff;
-            --color-muted: #a1a1aa;
-            --color-border: #27272a;
-            background: var(--color-bg);
-        }}
-
-        /* Density settings */
-        body.density-compact {{
-            --section-gap: 1.5rem;
-            --card-gap: 0.75rem;
-            --card-padding: 0.75rem;
-        }}
-        body.density-comfortable {{
-            --section-gap: 2.5rem;
-            --card-gap: 1.25rem;
-            --card-padding: 1.25rem;
-        }}
-        body.density-spacious {{
-            --section-gap: 4rem;
-            --card-gap: 2rem;
-            --card-padding: 1.75rem;
-        }}
-
-        /* View list mode */
-        body.view-list .stories-grid,
-        body.view-list .trend-grid {{
-            display: flex;
-            flex-direction: column;
-            gap: 0;
-        }}
-        body.view-list .story-card,
-        body.view-list .trend-card {{
-            background: transparent;
-            border: none;
-            border-bottom: 1px solid var(--color-border);
-            border-radius: 0;
-            padding: 0.5rem 0;
-            padding-left: 1.5rem;
-        }}
-        body.view-list .story-card img,
-        body.view-list .trend-card img,
-        body.view-list .card-image {{
-            display: none;
-        }}
-
-        {get_header_styles()}
-
-        /* Page Header */
-        .page-header {{
-            text-align: center;
-            padding: 4rem 2rem 3rem;
-            border-bottom: 1px solid var(--color-border);
-        }}
-
-        .page-title {{
-            font-family: var(--font-primary);
-            font-size: clamp(2rem, 5vw, 3.5rem);
-            font-weight: 700;
-            margin-bottom: 1rem;
-            background: linear-gradient(135deg, var(--color-accent) 0%, var(--color-accent-secondary) 100%);
-            -webkit-background-clip: text;
-            -webkit-text-fill-color: transparent;
-            background-clip: text;
-        }}
-
-        .page-subtitle {{
-            font-size: 1.1rem;
-            color: var(--color-muted);
-            max-width: 600px;
-            margin: 0 auto;
-        }}
-
-        /* Main Content */
-        .main-content {{
-            max-width: 1200px;
-            margin: 0 auto;
-            padding: 3rem 2rem;
-        }}
-
-        /* Media Sections */
-        .media-section {{
-            margin-bottom: 4rem;
-        }}
-
-        .section-header {{
-            display: flex;
-            align-items: center;
-            justify-content: space-between;
-            margin-bottom: 1.5rem;
-            padding-bottom: 1rem;
-            border-bottom: 1px solid var(--color-border);
-        }}
-
-        .section-title {{
-            font-family: var(--font-primary);
-            font-size: 1.5rem;
-            font-weight: 600;
-            display: flex;
-            align-items: center;
-            gap: 0.75rem;
-        }}
-
-        .section-icon {{
-            width: 32px;
-            height: 32px;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            background: var(--color-accent);
-            border-radius: 8px;
-        }}
-
-        .source-link {{
-            font-size: 0.85rem;
-            color: var(--color-accent);
-            text-decoration: none;
-            display: flex;
-            align-items: center;
-            gap: 0.5rem;
-            transition: color var(--transition);
-        }}
-
-        .source-link:hover {{
-            color: var(--color-accent-secondary);
-        }}
-
-        /* Image of the Day */
-        .image-container {{
-            background: var(--color-card-bg);
-            border: 1px solid var(--color-border);
-            border-radius: var(--radius);
-            overflow: hidden;
-        }}
-
-        .featured-image {{
-            width: 100%;
-            max-height: 70vh;
-            object-fit: contain;
-            background: #000;
-            display: block;
-        }}
-
-        .image-info {{
-            padding: 1.5rem;
-        }}
-
-        .media-title {{
-            font-family: 'Space Grotesk', sans-serif;
-            font-size: 1.3rem;
-            font-weight: 600;
-            margin-bottom: 0.75rem;
-        }}
-
-        .media-description {{
-            color: var(--color-muted);
-            font-size: 0.95rem;
-            line-height: 1.7;
-            margin-bottom: 1rem;
-        }}
-
-        .media-meta {{
-            display: flex;
-            flex-wrap: wrap;
-            gap: 1.5rem;
-            font-size: 0.85rem;
-            color: var(--color-muted);
-        }}
-
-        .meta-item {{
-            display: flex;
-            align-items: center;
-            gap: 0.5rem;
-        }}
-
-        .image-actions {{
-            display: flex;
-            gap: 1rem;
-            margin-top: 1rem;
-        }}
-
-        .action-btn {{
-            display: inline-flex;
-            align-items: center;
-            gap: 0.5rem;
-            padding: 0.75rem 1.25rem;
-            background: var(--color-accent);
-            color: #000;
-            font-weight: 600;
-            border-radius: var(--radius);
-            text-decoration: none;
-            transition: transform var(--transition), box-shadow var(--transition);
-        }}
-
-        .action-btn:hover {{
-            transform: translateY(-2px);
-            box-shadow: 0 10px 25px rgba(0,0,0,0.3);
-        }}
-
-        .action-btn.secondary {{
-            background: var(--color-card-bg);
-            color: var(--color-text);
-            border: 1px solid var(--color-border);
-        }}
-
-        /* Video of the Day */
-        .video-container {{
-            background: var(--color-card-bg);
-            border: 1px solid var(--color-border);
-            border-radius: var(--radius);
-            overflow: hidden;
-        }}
-
-        .video-embed {{
-            position: relative;
-            padding-bottom: 56.25%;
-            height: 0;
-            overflow: hidden;
-        }}
-
-        .video-embed iframe {{
-            position: absolute;
-            top: 0;
-            left: 0;
-            width: 100%;
-            height: 100%;
-            border: none;
-        }}
-
-        .video-info {{
-            padding: 1.5rem;
-        }}
-
-        .author-info {{
-            display: flex;
-            align-items: center;
-            gap: 0.75rem;
-            margin-top: 0.75rem;
-        }}
-
-        .author-avatar {{
-            width: 40px;
-            height: 40px;
-            border-radius: 50%;
-            background: var(--color-accent);
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            font-weight: 600;
-        }}
-
-        .author-name {{
-            color: var(--color-text);
-            text-decoration: none;
-            font-weight: 500;
-        }}
-
-        .author-name:hover {{
-            color: var(--color-accent);
-        }}
-
-        /* About Section */
-        .about-section {{
-            background: var(--color-card-bg);
-            border: 1px solid var(--color-border);
-            border-radius: var(--radius);
-            padding: 2rem;
-            margin-top: 3rem;
-        }}
-
-        .about-title {{
-            font-family: 'Space Grotesk', sans-serif;
-            font-size: 1.25rem;
-            font-weight: 600;
-            margin-bottom: 1rem;
-        }}
-
-        .about-text {{
-            color: var(--color-muted);
-            line-height: 1.7;
-        }}
-
-        .about-text a {{
-            color: var(--color-accent);
-            text-decoration: none;
-        }}
-
-        .about-text a:hover {{
-            text-decoration: underline;
-        }}
-
-        .source-list {{
-            display: grid;
-            grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
-            gap: 1.5rem;
-            margin-top: 1.5rem;
-        }}
-
-        .source-card {{
-            background: rgba(255,255,255,0.02);
-            border: 1px solid var(--color-border);
-            border-radius: 0.75rem;
-            padding: 1.25rem;
-        }}
-
-        .source-card h4 {{
-            font-size: 1rem;
-            margin-bottom: 0.5rem;
-        }}
-
-        .source-card p {{
-            font-size: 0.85rem;
-            color: var(--color-muted);
-        }}
-
-        {get_footer_styles()}
-
-        /* Mobile responsive */
-        @media (max-width: 768px) {{
-            .page-header {{
-                padding: 3rem 1rem 2rem;
-            }}
-
-            .main-content {{
-                padding: 2rem 1rem;
-            }}
-
-            .section-header {{
-                flex-direction: column;
-                align-items: flex-start;
-                gap: 0.75rem;
-            }}
-
-            .image-actions {{
-                flex-direction: column;
-            }}
-        }}
-    </style>
-</head>
-<body class="{base_mode}">
-    {build_header('media', date_str)}
-
-    <header class="page-header">
-        <h1 class="page-title">Media of the Day</h1>
-        <p class="page-subtitle">Curated daily content featuring stunning space imagery from NASA and award-winning short films from Vimeo's finest creators.</p>
-    </header>
-
-    <main class="main-content">
-        <!-- Image of the Day Section -->
-        {image_section}
-
-        <!-- Video of the Day Section -->
-        {video_section}
-
-        <!-- About Section -->
-        <div class="about-section">
-            <h3 class="about-title">About Media of the Day</h3>
-            <p class="about-text">
-                Every day, we curate the best visual content from trusted sources across the web.
-                Our selections feature stunning space imagery and thought-provoking short films
-                that inspire curiosity and creativity.
-            </p>
-            <div class="source-list">
-                <div class="source-card">
-                    <h4>🚀 NASA APOD</h4>
-                    <p>The Astronomy Picture of the Day features a different image or photograph of our universe each day, along with a brief explanation by a professional astronomer.</p>
-                </div>
-                <div class="source-card">
-                    <h4>🎬 Vimeo Staff Picks</h4>
-                    <p>Hand-picked by Vimeo's curation team, Staff Picks showcase the best short films, documentaries, and creative videos from filmmakers around the world.</p>
-                </div>
-            </div>
-        </div>
-    </main>
-
-    {build_footer(date_str)}
-
-    {get_theme_script()}
-</body>
-</html>"""
+        ok = generate_media_page(self.public_dir, self.media_data, design_data)
+        self.metrics.set_counter("media_page_generated", 1 if ok else 0)
+        if not ok:
+            logger.warning("Media page generation failed")
 
     def _step_generate_rss(self) -> None:
         """Generate RSS feed."""
